c# samples reorganized
