+++ lang/csharp/Continental.eCAL.Core.Samples/service/minimal_client_csharp.cs
