+++ lang/csharp/Continental.eCAL.Core.Samples/service/minimal_server_csharp.cs
